0.3.7

+++ app/src/main/java/com/example/music/music/presenter/SongDetPresenter.kt
@@ -21,7 +21,7 @@ class SongDetPresenter : BaseMvpPresenter<SongDetContract.IView, SongDetContract
         getModel().deldata(context,ids,playids)
     }
 
-    override fun delsongs(context: Context, data: Int, songids: Long) {
+    override fun delsongs(context: Context, data: Long, songids: Long) {
         getModel().delsongs(context,data,songids)
     }
